+++ src/features/edit-settings/index.ts
@@ -1,0 +1,4 @@
+export { LinksEditor } from "./LinksEditor";
+export { ProfileTab } from "./ProfileTab";
+export { SkillsTab } from "./SkillsTab";
+export { useSettingsForm } from "./useSettingsForm";