+++ src/shared/lib/index.ts
@@ -1,0 +1,10 @@
+export { downloadBlobFile, downloadTextFile } from "./download";
+export { LINK_ICONS, LINK_ICON_NAMES } from "./linkIcons";
+export { isMacLike, searchShortcutLabel } from "./platform";
+export { safeSetItem } from "./storage";
+export type { ProfileLink } from "./text";
+export { countryCode, excerpt, firstLine, hostLabel, obfuscateEmail, parseProfileLinks, stripMarkdown } from "./text";
+export { useScrollLock } from "./useScrollLock";
+export { cn, formatDateForInput, generateId } from "./utils";
+export type { ZipEntry } from "./zip";
+export { buildZip } from "./zip";